feat(apps): cms - upcoming semester virtual field.
