+++ apps/cms/routes/studios.ts
@@ -31,32 +31,12 @@ export async function getUpcomingSemesters(
 ) {
   const result = await context.query.Semester.findMany({
     query: `
+      id 
       name
       key
       studio {
           name
           key
-      }
-      initiatives
-      courseNumber
-      instructors {
-          name
-      }
-      previewThumbnail {
-        publicId
-      }
-      previewThumbAltText
-      previewSummary {
-          document
-      }
-      previewVideo {
-        file
-      }
-      captions {
-        url
-      }
-      previewVideoThumbnail {
-        publicId
       }
       `,
     where: { type: { equals: 'upcoming' } },
